Conditional fake_rpi requirement
