+++ runchrono.py
@@ -6,28 +6,10 @@ import time
 
 from argparse import ArgumentParser
 from PyQt5 import QtWidgets, QtCore
+from F3FChrono.Utils import is_running_on_pi
 import os.path
 
-def get_raspi_revision():
-    rev_file = '/sys/firmware/devicetree/base/model'
-    info = { 'pi': '', 'model': '', 'rev': ''}
-    raspi = model = revision = ''
-    try:
-        fd = os.open(rev_file, os.O_RDONLY)
-        line = os.read(fd,256)
-        os.close(fd)
-        print (line)
-        m=re.split(r'\s', line.decode('utf-8'))
-        if m:
-            info['pi'] = m[3]
-            info['model'] = m[5]
-        return info
-    except:
-        pass
-    return info
-
-pi=get_raspi_revision()
-if pi['pi'] == '':
+if not is_running_on_pi():
     # Replace libraries by fake ones
     import sys
     import fake_rpi
