+++ setup.py
@@ -1,8 +1,16 @@
 import setuptools
+from F3FChrono.Utils import is_running_on_pi
 
 with open("README.md", "r") as fh:
 
     long_description = fh.read()
+
+install_requires = [
+       'pandas', 'requests', 'pymysql', 'Django', 'pyserial', 'scipy', 'pyttsx3', 'netifaces'
+    ]
+
+if not is_running_on_pi():
+    install_requires.append('fake_rpi')
 
 setuptools.setup(
     name='F3FChrono',
@@ -20,8 +28,5 @@ setuptools.setup(
         "License :: GPL v3",
         "Operating System :: OS Independent",
      ],
-    install_requires=[
-        'pandas', 'requests', 'pymysql', 'Django', 'fake_rpi', 'pyserial', 'scipy', 'pyttsx3', 'netifaces'
-    ]
-
+    install_requires=install_requires
  )
